manejador_cloud/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.8
python-json-logger==2.0.7
gunicorn==22.0.0
python-dotenv==1.0.1
